Rename color_v2 and colorV2 to colorv2
