+++ ts/input/tex/AllPackages.ts
@@ -24,14 +24,14 @@
 import './base/BaseConfiguration.js';
 import './action/ActionConfiguration.js';
 import './ams/AmsConfiguration.js';
-import './ams_cd/AmsCdConfiguration.js';
+import './amscd/AmsCdConfiguration.js';
 import './bbox/BboxConfiguration.js';
 import './boldsymbol/BoldsymbolConfiguration.js';
 import './braket/BraketConfiguration.js';
 import './bussproofs/BussproofsConfiguration.js';
 import './cancel/CancelConfiguration.js';
 import './color/ColorConfiguration.js';
-import './color_v2/ColorV2Configuration.js';
+import './colorv2/ColorV2Configuration.js';
 import './config_macros/ConfigMacrosConfiguration.js';
 import './enclose/EncloseConfiguration.js';
 import './extpfeil/ExtpfeilConfiguration.js';
@@ -41,7 +41,7 @@ import './newcommand/NewcommandConfiguration.js';
 import './noerrors/NoErrorsConfiguration.js';
 import './noundefined/NoUndefinedConfiguration.js';
 import './physics/PhysicsConfiguration.js';
-import './tag_format/TagFormatConfiguration.js';
+import './tagformat/TagFormatConfiguration.js';
 import './unicode/UnicodeConfiguration.js';
 import './verb/VerbConfiguration.js';
